requirements.txt
treamlit
yfinance
pandas
pandas_ta
requests
lxml
html5lib
beautifulsoup4
